requirements.txt
django
requests
kafka-python
websocket-client
channels
channels_redis